Generalize import-flux2-assets.sh script
This is the first step towards supporting multiple versions of Flux to
be documented in parallel. The script has been condensed to only
require a single input parameter denoting the desired Flux major and
minor version for which to build the documentation.

This parameter is read from the environment variable `BRANCH` (which
Netlify) automatically provides and has to be either `main` or have
the format `vX-Y` which `X` denoting the major and `Y` denoting the
minor version desired to build the docs for.

The intent is to have branches following this naming pattern that
Netlify will directly deploy under the URL vX-Y.fluxcd.io.

The changed import-flux2-assets.sh script now conducts the following
steps in order to generate the docs for a specific version:

1. Find most recent patch version of Flux matching the given `BRANCH`.
2. For each controller fetch the `kustomization.yaml` from
   https://github.com/fluxcd/flux2/blob/FLUX_VERSION/manifests/bases/CONTROLLER/kustomization.yaml
3. From there fetch the `*.crds.yaml` file and the controller version.
4. For each CRD group version fetch the API docs from the controller
   repository as well as the spec docs.

Some special cases remain, e.g. for IRC/IAC which are supposed to be
documented under the same main navigation entry and for NC's `Event`
type which is not a CRD.

refs #1534

Signed-off-by: Max Jonas Werner <mail@makk.es>